==> res/datasets/SAMPLE_EasyClinic/Originals/Cleaning/Step 1. Create Base Oracles (R2D, D2C, R2C).ipynb <==
import pandas as pd
import numpy as np
import os
import sys

sys.path.append("../../../../Tracer/src")
from PreProcessing.Traces.TraceCreator import create_trace_matrix_from_trace_list

def create_trace_list(path_to_oracle):
    oracle = open(path_to_oracle).read().split("\n")
    trace_list = []
    for trace_item in oracle:
        if len(trace_item) > 0:
            source, targets = trace_item.split(":")
            for target in targets.split(" "):
                if len(target) != 0:
                    trace_list.append(" ".join([source, target]))
                else:
                    trace_list.append(" ".join([source, ""]))
    return create_trace_matrix_from_trace_list(trace_list)

# 1 -> 2

path_to_oracle = "../Originals/TracedMatrices/UC_ID.txt"
level_1_to_level_2 = create_trace_list(path_to_oracle)
print("shape: ", level_1_to_level_2.shape)
level_1_to_level_2.head(1)

level_1_to_level_2.columns

# 2 -> 3

path_to_oracle = "../Originals/TracedMatrices/ID_CC.txt"
level_2_to_level_3 = create_trace_list(path_to_oracle)
print("shape: ", level_2_to_level_3.shape)
level_2_to_level_3.head(1)

# 1 -> 3

path_to_oracle = "../Originals/TracedMatrices/UC_CC.txt"
level_1_to_level_3 = create_trace_list(path_to_oracle)
print("shape: ", level_1_to_level_3.shape)
level_1_to_level_3.head(1)

# Export

level_1_to_level_2.to_csv("../../Oracles/DataFrames/Level_1_to_Level_2.csv")
level_2_to_level_3.to_csv("../../Oracles/DataFrames/Level_2_to_Level_3.csv")
level_1_to_level_3.to_csv("../../Oracles/DataFrames/Level_1_to_Level_3.csv")
level_1_to_level_3.to_csv("../../Oracles/Relations.csv")
print("Done!")
