==> res/datasets/SAMPLE_EasyClinic/Originals/Cleaning/Step 2. Create Levels.ipynb <==
import pandas as pd
import numpy as np
import os
import sys

PATH_TO_SRC = "../../../../Tracer/src"

sys.path.append(PATH_TO_SRC)
from PreProcessing.Traces.TraceCreator import create_trace_matrix_from_trace_list
from PreProcessing.Levels.Cleaners import clean_doc
import warnings
warnings.filterwarnings(action='once')

def create_artifact(path_to_artifact_folder):
    artifact_files = list(filter(lambda f: f[0]!='.', os.listdir(path_to_artifact_folder)))
    identifiers = []
    docs = []
    for a_file in artifact_files:
        identifiers.append(a_file)
        file = open("/".join([path_to_artifact_folder, a_file]), encoding='utf-8', errors='ignore')
        doc = file.read()
        file.close()
        docs.append(clean_doc(doc))
    result = pd.DataFrame()
    result["id"] = identifiers
    result["text"] = docs
    return result

# Read Level

level_1 = create_artifact("../Originals/Definitions/Use_Cases")
level_1.head()

level_2 = create_artifact("../Originals/Definitions/Interaction_Diagrams")
level_2.head()

level_3 = create_artifact("../Originals/Definitions/Class_Descriptions")
level_3.head()

# Export

level_1.to_csv("../../Artifacts/Level_1.csv", index=False)
level_2.to_csv("../../Artifacts/Level_2.csv", index=False)
level_3.to_csv("../../Artifacts/Level_3.csv", index=False)
print("DOne!")
