==> res/datasets/SAMPLE_EasyClinic/Originals/Cleaning/Step 4. Export Intersection of Artifacts and Oracles.ipynb <==
import pandas as pd
import numpy as np

level_1 = pd.read_csv("../Level_1.csv")
print(len(level_1))

level_2 = pd.read_csv("../Level_2.csv")
print(len(level_2))

level_3 = pd.read_csv("../Level_3.csv")
print(len(level_3))

# Intersection

relations_df = pd.read_csv("../../Oracles/DataFrames/Level_1_to_Level_3.csv").set_index("id")
relations_df.head(1)

### Level 1 to Level 2

r_2_d = pd.read_csv("../../Oracles/DataFrames/Level_1_to_Level_2.csv").set_index("id")
r_2_d.head(1)

### Level 2 to Level 3

d_2_c = pd.read_csv("../../Oracles/DataFrames/Level_2_to_Level_3.csv").set_index("id")
d_2_c.head(1)

r_intersection = set(r_2_d.index) & set(relations_df.index) & set(level_1["id"])
d_intersection = set(r_2_d.columns) & set(d_2_c.index) & set(level_2["id"])
c_intersection = set(d_2_c.columns) & set(level_3["id"]) & set(relations_df.columns)

print("Number Requirements: ", len(r_intersection))
print("Number Designs: ", len(d_intersection))
print("Number Classes: ", len(c_intersection))

# Filter

relations_df = relations_df.loc[r_intersection][c_intersection]
r_2_d = r_2_d.loc[r_intersection][d_intersection]
d_2_c = d_2_c.loc[d_intersection][c_intersection]

level_1 = level_1.set_index("id").loc[r_intersection].reset_index()
level_2 = level_2.set_index("id").loc[d_intersection].reset_index()
level_3 = level_3.set_index("id").loc[c_intersection].reset_index()

assert relations_df.shape == (len(r_intersection), len(c_intersection))
assert r_2_d.shape == (len(r_intersection), len(d_intersection))
assert d_2_c.shape == (len(d_intersection), len(c_intersection))
assert len(level_1) == len(r_intersection)
assert len(level_2) == len(d_intersection)
assert len(level_3) == len(c_intersection)

# Export

r_2_d.sum().sum()

len(level_1) * len(level_2)

level_1.to_csv("../Level_1.csv")
level_2.to_csv("../Level_2.csv")
level_3.to_csv("../Level_3.csv")
np.save("../../Oracles/TracedMatrices/Level_1_to_Level_2.npy", r_2_d.values)
np.save("../../Oracles/TracedMatrices/Level_2_to_Level_3.npy", d_2_c.values)
relations_df.to_csv("../../Oracles/Relations.csv")
print("Done!")

a = r_2_d.values
b = d_2_c.values
similarity_matrix = np.dot(a, b)

similarity_matrix.shape

scores = np.apply_along_axis(max, 1, similarity_matrix)
print("Scores:", len(scores))
scores

len(level_1)
